# platform/openapi.json
{
  "components": {
    "schemas": {
      "Classification": {
        "description": "``confidence`` is always returned, so no caller is forced through our threshold.",
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "fallback": {
            "title": "Fallback",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "id",
          "topic",
          "confidence",
          "fallback"
        ],
        "title": "Classification",
        "type": "object"
      },
      "ClassifyItem": {
        "description": "``id`` is the caller's opaque correlation handle, echoed back unchanged.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "text"
        ],
        "title": "ClassifyItem",
        "type": "object"
      },
      "ClassifyRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClassifyItem"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "taxonomy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taxonomy Version"
          }
        },
        "required": [
          "items"
        ],
        "title": "ClassifyRequest",
        "type": "object"
      },
      "ClassifyResponse": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Errors",
            "type": "array"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Classification"
            },
            "title": "Results",
            "type": "array"
          },
          "taxonomy_version": {
            "title": "Taxonomy Version",
            "type": "string"
          }
        },
        "required": [
          "taxonomy_version",
          "results"
        ],
        "title": "ClassifyResponse",
        "type": "object"
      },
      "ErrorCode": {
        "description": "Locked. Adding a member changes the published contract.",
        "enum": [
          "invalid_request",
          "unsupported_taxonomy_version",
          "item_too_large",
          "rate_limited",
          "not_found",
          "internal"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "description": "One failure. ``item_id`` is set only when the failure concerns a single batch item.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "A request that failed as a whole. Per-item failures go in a response's ``errors``.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobAccepted": {
        "description": "The 202 body. ``job_id`` is a random public handle, not a storage key.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
      "JobState": {
        "description": "``results`` and ``errors`` are empty until the job reaches a terminal state.\n\nInputs are discarded at terminal state and the job itself 24 hours later; a poll after\nthat answers ``expired``, and later still ``404 not_found``.",
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Errors",
            "type": "array"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Summary"
            },
            "title": "Results",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobState",
        "type": "object"
      },
      "JobStatus": {
        "description": "Terminality is readable from this field alone; a caller polls until terminal.",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "partial",
          "failed",
          "expired"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "SourceDocument": {
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "source",
          "title",
          "text",
          "url"
        ],
        "title": "SourceDocument",
        "type": "object"
      },
      "SummarizeItem": {
        "description": "``id`` is the caller's opaque correlation handle, echoed back unchanged.",
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/SourceDocument"
            },
            "minItems": 1,
            "title": "Documents",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "documents"
        ],
        "title": "SummarizeItem",
        "type": "object"
      },
      "SummarizeRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SummarizeItem"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "max_sentences": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Sentences"
          },
          "style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style"
          }
        },
        "required": [
          "items"
        ],
        "title": "SummarizeRequest",
        "type": "object"
      },
      "SummarizeResponse": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Errors",
            "type": "array"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Summary"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "SummarizeResponse",
        "type": "object"
      },
      "Summary": {
        "description": "``summary`` is empty when ``withheld``. Branch on ``withheld``, not on the reason:\n``withhold_reason`` may grow values without a wire-major bump.",
        "properties": {
          "faithfulness_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Faithfulness Score",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "provenance": {
            "items": {
              "type": "string"
            },
            "title": "Provenance",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "withheld": {
            "title": "Withheld",
            "type": "boolean"
          },
          "withhold_reason": {
            "anyOf": [
              {
                "const": "below_faithfulness_bar",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withhold Reason"
          }
        },
        "required": [
          "id",
          "summary",
          "faithfulness_score",
          "withheld"
        ],
        "title": "Summary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Summarization & Classification Platform",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/classify": {
      "post": {
        "operationId": "classify_v1_classify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited; honour Retry-After."
          }
        },
        "summary": "Classify"
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobState"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited; honour Retry-After."
          }
        },
        "summary": "Job"
      }
    },
    "/v1/summarize": {
      "post": {
        "operationId": "summarize_v1_summarize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummarizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Batch above the sync ceiling."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited; honour Retry-After."
          }
        },
        "summary": "Summarize"
      }
    }
  }
}